ANY-TYPE! => ANY-VALUE! change
The `foo [any-type!]` name seems like it could be saying that foo can
be STRING! or INTEGER! or "any datatype".  ANY-VALUE! is clearer.
ANY-TYPE! is a synonym in <r3-legacy>.

Several of these cases will have to be reviewed when UNSET! is no
longer a valid "datatype" and you merely can indicate that a function
parameter is "optional".