#r2only tests to %old/r2-tests.r, lose #r3only flag
To bring better focus onto the work at hand with Ren/C testing, this
is the result of running a script to break out the Rebol2-based tests into
a separate file.

This file will not be very useful in terms of testing Rebol2 builds.  However,
it may inform the creation of tickets about missing Rebol3 features.

Consequently also removes the #r3 and #r3only flags from the main tests.