Check for non-standard spacing between functions.
A parse rule defines a standard for spacing between functions.  Source analysis code NON-STD-FUNC-SPACE is emitted if the new rule is failed.

Note: Source analysis parsing appears to be slower, probably due to the introduction of the BRACED rule to the parser which is needed to find the end of a function.