Unify rebol-to-c name comparison checks
There was a reverse transformation that would produce Rebol names
from a C string and check to make sure the names lined up when
checking the comment alignment with the source.  (e.g. the name of
a native with the C identifier inside of REBNATIVE(...)).

This gets rid of the reverse transformation, choosing instead to do the
same forward transformation of the Rebol word to the C string, to
reduce the total amount of code to keep in sync.