Copy local filename series before modifying
The `log-file-prefix: %r` line took a single copy of the series %r
and would modify it.  This would not be reset on subsequent calls to
run-recover, leading to a bug if it were called more than once in
the same session.

Modified for reasons of correctness (as well as the need to comply with
the new rule on the NewFunction branch that series in function bodies
are read-only).