Fix APPLY EVAL test, simple EVAL/ONLY tests
This adds two very simple EVAL/ONLY tests.

In a very appropriate demonstration of the breakages that EVAL (and
refinement revocation) are here to fix:

    apply :eval [:add 4 4]

EVAL had no refinements, then it got /ONLY so this has to become

    apply :eval [:add false 4 4]

The breaking is bad, the resulting callsite is uninformative, and it
is a canon example of why APPLY as written is no longer permitted in
the core of the language (but written as a bad legacy mezzanine).