Remove literal #[closure!] test
CLOSURE! no longer exists as an actual native datatype distinct from
function.  The literal encoding of closures in construction syntax
should ultimately be a matter of marking parameters as being durable
or non-durable, which is speculated to be a typeset property literally
signified by #[true] or #[false] (to use a type cue vs. introducing
a special-purpose keyword.