Update test assuming x and y unset
The %core-tests.r don't have an insulation of each test running in its
own isolated context.  One test was assuming failure on the idea that
x was not defined, so that when a test did involve setting an x...it
failed.

This unsets as a workaround, but ideally the test driver would create
a new context for each test.