De-tab pending tests, migrate EVAL apply
The test that once applied apply eventually turned into a mess of
taking the legacy apply and applying it on the now-variadic eval
function...which is not currently an agenda item to support in the
legacy apply.

Moving it here in case a positional apply that needs to work on EVAL
does become a priority, as a reminder of the case.